pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pensions-panorama"
version = "0.1.0"
description = "Comparative pension dataset and country briefs generator (Pensions-at-a-Glance style)"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Pensions Panorama" }]

dependencies = [
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "typer[all]>=0.9.0",
    "requests>=2.31.0",
    "pandas>=2.1.0",
    "numpy>=1.26.0",
    "openpyxl>=3.1.2",
    "matplotlib>=3.8.0",
    "PyYAML>=6.0.1",
    "requests-cache>=1.1.1",
    "diskcache>=5.6.3",
    "pyarrow>=14.0.0",
    "tenacity>=8.2.3",
    "rich>=13.7.0",
    "jinja2>=3.1.2",
    "python-dateutil>=2.8.2",
    "tabulate>=0.9.0",
    "streamlit>=1.28.0",
    "plotly>=5.18.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "responses>=0.24.1",
    "black>=23.11.0",
    "ruff>=0.1.6",
    "mypy>=1.7.0",
    "types-PyYAML>=6.0.12",
    "types-requests>=2.31.0",
]

[project.scripts]
pp = "pensions_panorama.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["pensions_panorama*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true
